changing the way dashboard modules behave as disscussed on the list. You now populate your available module with the dash_modules filter.
Also changing the way the dash_module_* filter behaves. Now you filter 
the $modules array instead of just the content, which allows you to 
change the title, content, and options keys. This method also simplifies 
the template used to generate the content of the modules. See the 
coredashmodules plugin for examples..



git-svn-id: https://svn.habariproject.org/habari/trunk/htdocs/system@2017 653ae4dd-d31e-0410-96ef-6bf7bf53c507